mini_projects/security-stack/secret-managment-in-k8s/implementing/backend-service/application-container/requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=3.0.0
python-dotenv>=1.0.0
PyMySQL>=1.0.2
cryptography>=41.0.0
